Add correlation id in preflight
