+++ node-runner-cli/monitoring/__init__.py
@@ -1,0 +1,74 @@
+import requests
+import os
+import os.path
+from pathlib import Path
+
+from env_vars import NODE_END_POINT
+from utils.utils import Helpers, run_shell_command
+
+
+class Monitoring():
+
+    @staticmethod
+    def setup_prometheus_yml(default_prometheus_yaml_url):
+        req = requests.Request('GET', f'{default_prometheus_yaml_url}')
+        prepared = req.prepare()
+
+        resp = Helpers.send_request(prepared, print_response=False)
+        Path("monitoring/prometheus").mkdir(parents=True, exist_ok=True)
+        with open("monitoring/prometheus/prometheus.yml", 'wb') as f:
+            f.write(resp.content)
+
+    @staticmethod
+    def setup_datasource(default_datasource_cfg_url):
+        req = requests.Request('GET', f'{default_datasource_cfg_url}')
+        prepared = req.prepare()
+        resp = Helpers.send_request(prepared, print_response=False)
+        Path("monitoring/grafana/provisioning/datasources").mkdir(parents=True, exist_ok=True)
+        with open("monitoring/grafana/provisioning/datasources/datasource.yml", 'wb') as f:
+            f.write(resp.content)
+
+    @staticmethod
+    def setup_dashboard(default_dashboard_cfg_url, files):
+        for file in files:
+            req = requests.Request('GET', f'{default_dashboard_cfg_url}/{file}')
+            prepared = req.prepare()
+            resp = Helpers.send_request(prepared, print_response=False)
+            Path("monitoring/grafana/provisioning/dashboards").mkdir(parents=True, exist_ok=True)
+            with open(f"monitoring/grafana/provisioning/dashboards/{file}", 'wb') as f:
+                f.write(resp.content)
+
+    @staticmethod
+    def setup_external_volumes():
+        run_shell_command(["docker", "volume", "create", "prometheus_tsdb"])
+        run_shell_command(["docker", "volume", "create", "grafana-storage"])
+
+    @staticmethod
+    def setup_monitoring_containers(default_monitoring_cfg_url):
+        req = requests.Request('GET', f'{default_monitoring_cfg_url}')
+        prepared = req.prepare()
+        resp = Helpers.send_request(prepared, print_response=False)
+        Path("monitoring").mkdir(parents=True, exist_ok=True)
+        with open("monitoring/node-monitoring.yml", 'wb') as f:
+            f.write(resp.content)
+
+    @staticmethod
+    def start_monitoring(composefile):
+        user = Helpers.get_nginx_user(usertype="metrics", default_username="metrics")
+        if os.environ.get('%s' % NODE_END_POINT) is None:
+            print(
+                f"{NODE_END_POINT} environment not setup. Fetching the IP of node assuming the monitoring is run on the same machine machine as "
+                "the node.")
+            ip = Helpers.get_public_ip()
+            node_endpoint = f"https://{ip}"
+        else:
+            node_endpoint = os.environ.get(NODE_END_POINT)
+        run_shell_command(f'docker-compose -f {composefile} up -d',
+                          env={
+                              "BASIC_AUTH_USER_CREDENTIALS": f'{user["name"]}:{user["password"]}',
+                              f"{NODE_END_POINT}": node_endpoint
+                          }, shell=True)
+
+    @staticmethod
+    def stop_monitoring(composefile, remove_volumes):
+        Helpers.docker_compose_down(composefile, remove_volumes)
